+++ example/lib/main.dart
@@ -1,8 +1,10 @@
-import 'package:example/core/packages.dart';
 import 'package:flutter/material.dart';
+import 'package:example/core/packages.dart';
+
+import 'package:example/models/models.module.dart';
+import 'core/services/firebase.service.dart';
 
 import 'app/app.view.dart';
-import 'core/services/firebase.service.dart';
 
 void main() => runApp(const Root());
 
@@ -13,11 +15,7 @@ class Root extends StatelessWidget {
   Widget build(BuildContext context) {
     return MultiProvider(
       providers: [
-        StreamProvider<FirebaseUser>.value(value: AuthService.user),
-        StreamProvider<AuthStatus>.value(
-          value: AuthService.authStatus,
-          initialData: AuthStatus.NOT_DONE,
-        ),
+        StreamProvider<Profile>.value(value: authService.profile),
       ],
       child: MyApp(),
     );