+++ lib/pmvvm.dart
@@ -2,6 +2,9 @@ library pmvvm;
 
 export 'views/hook.view.dart';
 export 'views/stateless.view.dart';
+
+export 'utils/extensions.dart';
+
 export 'mvvm-builder.widget.dart';
 export 'mvvm.widget.dart';
 export 'view-model.dart';