+++ lib/views/hook.view.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:provider/provider.dart';
+import 'package:pmvvm/pmvvm.dart';
 
 abstract class HookView<T> extends HookWidget {
   const HookView({Key key, this.reactive = true});